requirements.txt
fastapi[standard]
jinja2
aiofiles
openai
sqlalchemy
python-docx
htmldocx
passlib
bcrypt
python-dotenv